Commit 4 - Organisation, adding combat anims and State machine camera rotation
Organised the project moveing anims and envrionment asset to the game data folder and keeping system packages to the base asset directory
Downloaded and imported the combat anims to design the full control scheme
Added the camera rotation function to the state machine to provide the functionality only when in required states like moving. This gives the benefits of a nice and steady camera when walking and the benefits of an orbital camera when idle to get a look at your character.