Commit 7 - Changed state machine and added partial jumping functionality
Changed the state machine to have more of the functionality at a base level but this defeats the purpose of seperating the code into different states as the base functions are still there combined with the update function mean some functionailty happens accross states which is defeating the purpose of a state machine. Main problem is movement in a state that has no movement function.

Added the ability to jump that transitions into the falling anim but not the falling stat. Need to add the ability to land and transiton into the falling state from both jumping and walking/running/sprinting off a ledge. Also there is a bug where the jumping force isn't applied when idle.