Commit 8 - Fixed Jump Bug and State Machine organisation
Fixed the idle jumping bug, the cause was that when no input of movement was detected, the character move command wasn't run resulting in the jump force not being applied. this is not applied in the jump states movement as an else addition instead of retruning to Idle on a no input movement as the movement isn't tied to states is the act of landing that is tied to changing states

Got rid of all the functions being in the base functionlaity and returned the state mahcine to have seperated slightly different functionality for each funtion with only globally applied functionlity remaining in the base liek the Camera Position Follow and Gravity.